Improves packages organization
Creates application layer packages (rest, business, persistence)
in each domain package (topic, poll, vote).

Breaks the monolithic integration test in two,
one for each controller.
