--- src/main/java/com/vollino/poll/service/exception/ClientErrorException.java
@@ -1,15 +1,0 @@
-package com.vollino.poll.service.exception;
-
-/**
- * @author Bruno Vollino
- */
-public class ClientErrorException extends RuntimeException {
-
-    public ClientErrorException(String message) {
-        super(message);
-    }
-
-    public ClientErrorException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
