--- src/main/java/com/vollino/poll/service/exception/DataIntegrityException.java
@@ -1,15 +1,0 @@
-package com.vollino.poll.service.exception;
-
-/**
- * @author Bruno Vollino
- */
-public class DataIntegrityException extends ClientErrorException {
-
-    public DataIntegrityException(String message) {
-        super(message);
-    }
-
-    public DataIntegrityException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
