+++ src/main/java/com/vollino/poll/service/exception/business/ClientErrorException.java
@@ -1,0 +1,15 @@
+package com.vollino.poll.service.exception.business;
+
+/**
+ * @author Bruno Vollino
+ */
+public class ClientErrorException extends RuntimeException {
+
+    public ClientErrorException(String message) {
+        super(message);
+    }
+
+    public ClientErrorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
