+++ src/main/java/com/vollino/poll/service/exception/business/DataIntegrityException.java
@@ -1,0 +1,15 @@
+package com.vollino.poll.service.exception.business;
+
+/**
+ * @author Bruno Vollino
+ */
+public class DataIntegrityException extends ClientErrorException {
+
+    public DataIntegrityException(String message) {
+        super(message);
+    }
+
+    public DataIntegrityException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
