+++ src/main/java/com/vollino/poll/service/poll/business/Clock.java
@@ -1,0 +1,16 @@
+package com.vollino.poll.service.poll.business;
+
+import org.springframework.stereotype.Component;
+
+import java.time.ZonedDateTime;
+
+/**
+ * @author Bruno Vollino
+ */
+@Component
+public class Clock {
+
+    public ZonedDateTime now() {
+        return ZonedDateTime.now();
+    }
+}
