--- src/main/java/com/vollino/poll/service/poll/vote/VoteRepository.java
@@ -1,11 +1,0 @@
-package com.vollino.poll.service.poll.vote;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * @author Bruno Vollino
- */
-@Repository
-public interface VoteRepository extends CrudRepository<Vote, VoteId> {
-}
