--- src/main/java/com/vollino/poll/service/topic/Topic.java
@@ -1,68 +1,0 @@
-package com.vollino.poll.service.topic;
-
-import com.google.common.base.MoreObjects;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import java.util.Objects;
-
-/**
- * @author Bruno Vollino
- */
-@Entity
-@Table(name = "topic")
-public class Topic {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @NotBlank(message = "description is mandatory")
-    private String description;
-
-    public Topic() {
-    }
-
-    public Topic(Long id, String description) {
-        this.id = id;
-        this.description = description;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Topic topic = (Topic) o;
-        return Objects.equals(getId(), topic.getId()) &&
-                Objects.equals(getDescription(), topic.getDescription());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getId(), getDescription());
-    }
-
-    @Override
-    public String toString() {
-        return MoreObjects.toStringHelper(this)
-                .add("id", id)
-                .add("description", description)
-                .toString();
-    }
-}
