--- src/main/java/com/vollino/poll/service/topic/TopicRepository.java
@@ -1,11 +1,0 @@
-package com.vollino.poll.service.topic;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * @author Bruno Vollino
- */
-@Repository
-public interface TopicRepository extends CrudRepository<Topic, Long> {
-}
